+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## next
  * ui-router declarations improvements
+ * bugfix duplicated injects array in case of redundant /* @ngInject */
 
 ## v0.9.3 2014-05-16
  * /* @ngInject */ object literal support